# zstd: Fast copy from literals in asm routine