@WojciechMula 2022-04-24
@klauspost  I created checked this because of the comment in `gen.go`: "Investigate if it is possible to consistently overallocate literals." As far I understand `blockdec.decodeLiterals`, there are quite a lot of places where we allocate literals buffers or we grab some slices from other buffers. Thus I tried something easier as a starter. :)